% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codestocolors.R
\name{codes_to_colors}
\alias{codes_to_colors}
\title{Converting numeric ID codes to listed color name codes}
\usage{
codes_to_colors(codes, available.colors = NULL)
}
\arguments{
\item{total_length}{the number of unique positions to be marked on your animal. (This can be thought of as the total number of positions on which color bands or paint marks will be applied.)}

\item{redundancy}{the number of erasures that can occur without disrupting unique identification. This determines how robust the scheme is.}

\item{alphabet}{an integer representing the 'alphabet size.' This is the number of unique marks (think different paint colors or varieties of bands) at your disposal. Note: unlike the Reed-Solomon inspired function, \code{\link{reed_solomon}}, this function can take non-prime values.}

\item{num.tries}{the number of iterations that will be run before choosing the best option. Increasing this number increases the running time.}
}
\value{
a list of unique color codes that fit the provided parameters.
}
\description{
This is a helper function
}
\note{
HEY THIS IS A NOTE
}
\examples{
plot_crayons()

}
\references{
\url{http://DOI OF THE PAPER} for the orginal source
}
\seealso{
\code{\link{birdtag}} for an example extension based on leg-banding birds,
}
\author{
Andrew Burchill, \email{andrew.burchill@asu.edu}
}
\keyword{reed_solomon}
