% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{rabi}
\alias{rabi}
\alias{rabi-package}
\title{rabi: a package for generating Robust Animal Behavior IDs}
\description{
This is a mostly empty packagesss to learn roxygen documentation.
}
\details{
Hello allows me to learn how to write documentation in comment blocks
co-located with code.
}
\section{Foo functions}{

The foo functions ...
}

\references{
LINK TO MY PAPER
}
\author{
Andrew Burchill: \email{andrew.burchill@asu.edu}
}
