% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poversample.R
\name{reed_solomon}
\alias{reed_solomon}
\title{Polynomial color coding scheme generator}
\usage{
reed_solomon(total_length, redundancy, alphabet, available.colors = NULL)
}
\arguments{
\item{total_length}{the number of unique positions to be marked on the animal. (This can be thought of as the total number of positions on which color bands or paint marks will be applied.) Note: Reed-Solomon coding requires the total length of the ID to be less than or equal to the value of \code{alphabet}.}

\item{redundancy}{the number of erasures that can occur without disrupting surety of unique identification. This value determines how robust the scheme is to erasures.}

\item{alphabet}{an integer representing the 'alphabet size.' This is the number of unique markings (think different paint colors, symbols, or varieties of bands) at your disposal. Note: Reed-Solomon coding requires this value to be a prime number. If a non-prime is entered, the function will automatically adjust it to the nearest previous prime.}

\item{available.colors}{an optional list of strings that contains the names of the unique markings which compose the given 'alphabet' (e.g. "blue", "red", "yellow", etc.). If left blank, the mapping can be done at any later time using \code{\link{codes_to_colors}}. Additionally, the length of this list must match the 'alphabet size' given above.}
}
\value{
a list containing  with all the possible unique identities included
}
\description{
Creates a color coding scheme used to mark and identify individual animals using Reed-Solomon error-correction codes. The codes are robust to an arbitrary number of color-slot erasures.
}
\examples{
plot_crayons()

}
\references{
\url{http://DOI OF THE PAPER} for the orginal source
}
\seealso{
\code{\link{birdtag}} for an example extension based on leg-banding birds,
}
\author{
Andrew Burchill, \email{andrew.burchill@asu.edu}
}
\keyword{reed_solomon}
