% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplecodes.R
\name{simple_code}
\alias{simple_code}
\title{Simple color coding scheme generator}
\usage{
simple_code(total_length, alphabet, available.colors = NULL)
}
\arguments{
\item{total_length}{the number of unique positions to be marked on the animal. (This can be thought of as the total number of positions on which color bands or paint marks will be applied.)}

\item{alphabet}{an integer representing the 'alphabet size.' This is the number of unique markings (think different paint colors, symbols, or varieties of bands) at your disposal.}

\item{available.colors}{an optional list of strings that contains the names of the unique markings which compose the given 'alphabet' (e.g. "blue", "red", "yellow", etc.). If left blank, the mapping can be done at any later time using \code{\link{codes_to_colors}}. Additionally, the length of this list must match the 'alphabet size' given above.}
}
\value{
a dataframe with all the possible unique identities included
}
\description{
Creates a simple color coding scheme used to mark and identify individual animals. Even if one marking is removed, the entire code can be reconstructed.
}
\examples{
plot_crayons()

}
\references{
\url{http://en.wikipedia.org/wiki/List_of_Crayola_crayon_colors}
}
\seealso{
\code{\link{brocolors}}
}
\author{
Karl W Broman, \email{kbroman@biostat.wisc.edu}
}
\keyword{hplot}
